bst/requirements.txt
django
django-crispy-forms
mysqlclient
torch
pillow
numpy
opencv-python
pandas
requests
pyyaml
tqdm
torchvision
matplotlib
seaborn